refactor: 重构模型映射逻辑及相关组件集成
- 将 ModelMapping 相关功能，从 EditChannel 中解耦
- 新增 ModelMappingEditor.js modelMappingUtils.js useModelMapping.js
- 在 EditTagModal 中增加 ModelMappingEditor 组件，并引入 useModelMapping
